requirements.txt
click
requests
python-dotenv
boto3
tqdm


